# static/openapi.yaml
openapi: 3.0.3
info:
  title: SodhiCable MES v4.0 API
  description: |
    REST API for the SodhiCable Manufacturing Execution System.

    Aligned with **MESA-11** (Manufacturing Enterprise Solutions Association) functional
    model and the **ISA-95** (IEC 62264) enterprise-control integration standard.

    The API spans ISA-95 Levels 1-4:
    - **Level 1-2** - Process control, sensors, equipment monitoring
    - **Level 3** - MES core: work orders, scheduling, quality, traceability, performance, labor
    - **Level 4** - ERP integration: S&OP, purchase orders, order-to-cash, ISA-95 messaging
    - **Cross-Level** - AI/ML intelligence layer (proposed MESA F12)

    All responses use JSON. Paginated endpoints return an envelope with
    `{data, page, per_page, total}`.
  version: 4.0.0
  contact:
    name: Stephen Eacuello
    email: seacuello@uri.edu
  license:
    name: Academic Use
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/traceability/certificate/{wo_id}:
    get:
      tags: ["Level 3 - F10 Traceability"]
      summary: Certificate of compliance
      description: |
        Generates a structured certificate of compliance for a work order,
        aggregating product details, test results, lot genealogy, and
        applicable specification (MIL-DTL-24643, IEEE 1580, UL 2196, etc.).
      operationId: getCertificate
      parameters:
        - name: wo_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate data
          content:
            application/json:
              schema:
                type: object
                properties:
                  wo_id:
                    type: string
                  product:
                    type: object
                    properties:
                      product_id:
                        type: string
                      name:
                        type: string
                      family:
                        type: string
                      awg:
                        type: string
                  customer:
                    type: string
                    nullable: true
                  order_qty_kft:
                    type: number
                  test_summary:
                    type: array
                    items:
                      type: object
                  material_traceability:
                    type: array
                    items:
                      type: object
                  compliance_spec:
                    type: string
                  issue_date:
                    type: string
                    format: date-time
                  all_tests_passed:
                    type: boolean
        "404":
          description: Work order not found
  /api/traceability/trace:
    get:
      tags: ["Level 3 - F10 Traceability"]
      summary: Bi-directional lot trace
      description: |
        Performs a forward or backward recursive trace through the lot genealogy.
        Forward trace finds all children (downstream products); backward trace
        finds all parent materials. Max depth: 10 levels.
      operationId: traceLot
      parameters:
        - name: lot
          in: query
          required: true
          schema:
            type: string
          description: Lot number to trace
        - name: direction
          in: query
          schema:
            type: string
            enum: [forward, backward]
            default: forward
      responses:
        "200":
          description: Trace results
          content:
            application/json:
              schema:
                type: object
                properties:
                  lot:
                    type: string
                  direction:
                    type: string
                  trace:
                    type: array
                    items:
                      type: object

  # ====================================================================
  # Level 2 - Equipment
  # ====================================================================
...